ctr-gcn-extended/requirements.txt
absl-py
aiocontextvars
asn1crypto
awscli
backcall
bleach
boto
botocore
cached-property
cachetools
certifi
cffi
chardet
cloudpickle
colorama
conda-pack
contextvars
cryptography
cycler
Cython
dask
dataclasses
decorator
docutils
easydict
entrypoints
future
fvcore
glog
google-auth
google-auth-oauthlib
grpcio
h5py
idna
imageio
immutables
importlib-metadata
iopath
ipykernel
ipython
ipython-genutils
ipywidgets
jedi
Jinja2
jmespath
joblib
jsonschema
jupyter
jupyter-client
jupyter-console
jupyter-core
kiwisolver
loguru
Markdown
MarkupSafe
matplotlib
mistune
mkl-fft
mkl-random
mkl-service
msgpack
msgpack-numpy
multimethod
nbconvert
nbformat
networkx
notebook
numpy
oauthlib
opencv-python
pandocfilters
parso
pexpect
pickleshare
Pillow
Pillow-SIMD
portalocker
prometheus-client
prompt-toolkit
protobuf
psutil
ptyprocess
pyasn1
pyasn1-modules
Pygments
pyOpenSSL
pyparsing
PySocks
python-dateutil
python-gflags
PyWavelets
PyYAML
pyzmq
qtconsole
requests
requests-oauthlib
rsa
ruamel.yaml
s3transfer
scikit-image
scikit-learn
scipy
Send2Trash
tabulate
tensorboard
tensorboard-plugin-wit
tensorboardX
tensorpack
termcolor
terminado
testpath
thop
threadpoolctl
toolz
torch
torchpack
torchvision
tornado
tqdm
traitlets
typing-extensions
urllib3
wcwidth
webencodings
Werkzeug
widgetsnbextension
yacs
zipp
